procd: Allow setting user and/or group wifi-presence runs as
This addresses an issue reported in #3 where the hostapd socket files
are owned by a different user/group than wifi-presence is being run as.
An error occurs at wifi-presence startup because of lack of permissions
accessing the sockets.
Likely caused by use of seccomp or similar, triggering hostapd to be run
as 'network' user/group [0].

This adds the ability to configure the user and/or group that runs the
wifi-presence process in the config file.

[0]: https://github.com/openwrt/openwrt/blob/ec6293febc244d187e71a6e54f44920be679cde4/package/network/services/hostapd/files/wpad.init#L35-L36